+++ test/unit/fuzz-tools.test.js
@@ -350,6 +350,7 @@ import * as hartofilakidisDdhV370 from '../../lib/hartofilakidis-ddh-v370.js';
 import * as cRadsV371 from '../../lib/c-rads-v371.js';
 import * as cadRadsV372 from '../../lib/cad-rads-v372.js';
 import * as niRadsV373 from '../../lib/ni-rads-v373.js';
+import * as pauwelsFemoralNeckV374 from '../../lib/pauwels-femoral-neck-v374.js';
 import * as tbTesting from '../../lib/tb-testing.js';
 
 const MODULES = {
@@ -645,6 +646,7 @@ const MODULES = {
   'c-rads-v371.js': cRadsV371,
   'cad-rads-v372.js': cadRadsV372,
   'ni-rads-v373.js': niRadsV373,
+  'pauwels-femoral-neck-v374.js': pauwelsFemoralNeckV374,
   'tb-testing.js': tbTesting,
 };